Using yarn, recompiled contracts

+++ contracts/testing/CentralizedArbitrator.sol
@@ -15,34 +15,36 @@ import "../arbitration/Arbitrator.sol";
  *  @dev This is a centralized arbitrator deciding alone on the result of disputes. No appeals are possible.
  */
 contract CentralizedArbitrator is Arbitrator {
-
     address public owner = msg.sender;
-    uint arbitrationPrice; // Not public because arbitrationCost already acts as an accessor.
-    uint constant NOT_PAYABLE_VALUE = (2**256-2)/2; // High value to be sure that the appeal is too expensive.
+    uint256 arbitrationPrice; // Not public because arbitrationCost already acts as an accessor.
+    uint256 constant NOT_PAYABLE_VALUE = (2**256 - 2) / 2; // High value to be sure that the appeal is too expensive.
 
     struct DisputeStruct {
         Arbitrable arbitrated;
-        uint choices;
-        uint fee;
-        uint ruling;
+        uint256 choices;
+        uint256 fee;
+        uint256 ruling;
         DisputeStatus status;
     }
 
-    modifier onlyOwner {require(msg.sender==owner, "Can only be called by the owner."); _;}
+    modifier onlyOwner {
+        require(msg.sender == owner, "Can only be called by the owner.");
+        _;
+    }
 
     DisputeStruct[] public disputes;
 
     /** @dev Constructor. Set the initial arbitration price.
      *  @param _arbitrationPrice Amount to be paid for arbitration.
      */
-    constructor(uint _arbitrationPrice) public {
+    constructor(uint256 _arbitrationPrice) public {
         arbitrationPrice = _arbitrationPrice;
     }
 
     /** @dev Set the arbitration price. Only callable by the owner.
      *  @param _arbitrationPrice Amount to be paid for arbitration.
      */
-    function setArbitrationPrice(uint _arbitrationPrice) public onlyOwner {
+    function setArbitrationPrice(uint256 _arbitrationPrice) public onlyOwner {
         arbitrationPrice = _arbitrationPrice;
     }
 
@@ -50,7 +52,12 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _extraData Not used by this contract.
      *  @return fee Amount to be paid.
      */
-    function arbitrationCost(bytes memory _extraData) public override view returns(uint fee) {
+    function arbitrationCost(bytes memory _extraData)
+        public
+        override
+        view
+        returns (uint256 fee)
+    {
         return arbitrationPrice;
     }
 
@@ -59,7 +66,12 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _extraData Not used by this contract.
      *  @return fee Amount to be paid.
      */
-    function appealCost(uint _disputeID, bytes memory _extraData) public override view returns(uint fee) {
+    function appealCost(uint256 _disputeID, bytes memory _extraData)
+        public
+        override
+        view
+        returns (uint256 fee)
+    {
         return NOT_PAYABLE_VALUE;
     }
 
@@ -69,15 +81,22 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _extraData Can be used to give additional info on the dispute to be created.
      *  @return disputeID ID of the dispute created.
      */
-    function createDispute(uint _choices, bytes memory _extraData) public override payable returns(uint disputeID)  {
+    function createDispute(uint256 _choices, bytes memory _extraData)
+        public
+        override
+        payable
+        returns (uint256 disputeID)
+    {
         super.createDispute(_choices, _extraData);
-        disputes.push(DisputeStruct({
-            arbitrated: Arbitrable(msg.sender),
-            choices: _choices,
-            fee: msg.value,
-            ruling: 0,
-            status: DisputeStatus.Waiting
-            }));
+        disputes.push(
+            DisputeStruct({
+                arbitrated: Arbitrable(msg.sender),
+                choices: _choices,
+                fee: msg.value,
+                ruling: 0,
+                status: DisputeStatus.Waiting
+            })
+        );
         disputeID = disputes.length - 1; // Create the dispute and return its number.
         emit DisputeCreation(disputeID, Arbitrable(msg.sender));
     }
@@ -86,23 +105,26 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _disputeID ID of the dispute to rule.
      *  @param _ruling Ruling given by the arbitrator. Note that 0 means "Not able/wanting to make a decision".
      */
-    function _giveRuling(uint _disputeID, uint _ruling) internal {
+    function _giveRuling(uint256 _disputeID, uint256 _ruling) internal {
         DisputeStruct storage dispute = disputes[_disputeID];
         require(_ruling <= dispute.choices, "Invalid ruling.");
-        require(dispute.status != DisputeStatus.Solved, "The dispute must not be solved already.");
+        require(
+            dispute.status != DisputeStatus.Solved,
+            "The dispute must not be solved already."
+        );
 
         dispute.ruling = _ruling;
         dispute.status = DisputeStatus.Solved;
         /* solium-disable-next-line */
         msg.sender.send(dispute.fee); // Avoid blocking.
-        dispute.arbitrated.rule(_disputeID,_ruling);
+        dispute.arbitrated.rule(_disputeID, _ruling);
     }
 
     /** @dev Give a ruling. UNTRUSTED.
      *  @param _disputeID ID of the dispute to rule.
      *  @param _ruling Ruling given by the arbitrator. Note that 0 means "Not able/wanting to make a decision".
      */
-    function giveRuling(uint _disputeID, uint _ruling) public onlyOwner {
+    function giveRuling(uint256 _disputeID, uint256 _ruling) public onlyOwner {
         return _giveRuling(_disputeID, _ruling);
     }
 
@@ -110,7 +132,12 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _disputeID ID of the dispute to rule.
      *  @return status The status of the dispute.
      */
-    function disputeStatus(uint _disputeID) public override view returns(DisputeStatus status) {
+    function disputeStatus(uint256 _disputeID)
+        public
+        override
+        view
+        returns (DisputeStatus status)
+    {
         return disputes[_disputeID].status;
     }
 
@@ -118,7 +145,12 @@ contract CentralizedArbitrator is Arbitrator {
      *  @param _disputeID ID of the dispute to rule.
      *  @return ruling The ruling which would or has been given.
      */
-    function currentRuling(uint _disputeID) public override view returns(uint ruling) {
+    function currentRuling(uint256 _disputeID)
+        public
+        override
+        view
+        returns (uint256 ruling)
+    {
         return disputes[_disputeID].ruling;
     }
 }
